refactor: criação da pasta api/ para api.js
